Backend/requirements.txt
flask
flask-restful
PyJWT
SQLAlchemy
sqlalchemy-pytds
gunicorn